cloudrun/db-service/requirements.txt
Flask==2.2.2
SQLAlchemy==2.0.39
pg8000==1.31.2
google-cloud-alloydb-connector==1.7.0
functions-framework==3.*
Werkzeug==2.3.7
google-cloud-secret-manager==2.21.0
Faker==0.7.4
